Clarify that Eduroam is considered off campus
We'd already make a note at the higher level, but this adds a note to the top
of the on campus instructions pointing out that these instructions won't work
if you're using Eduroam.  There's a supplemental note that being connected to
the VPN does give you on campus status as far as these instructions are
concerned.